refactor edge columns as separate plugin
