+++ index.js
@@ -32,5 +32,6 @@ module.exports = function PgManyToManyPlugin(builder, options) {
   });
 
   require("./src/PgManyToManyRelationPlugin.js")(builder, options);
+  require("./src/PgManyToManyRelationEdgeColumnsPlugin.js")(builder, options);
   require("./src/PgManyToManyRelationEdgeTablePlugin.js")(builder, options);
 };
